refactor: rename lib/ to src/ across all packages
Address PR #28 review feedback from @kennyek to use conventional src/
directory for TypeScript source files instead of lib/.

Renamed in: tapeworm, mongodb, indexeddb, remote stores.
Hybrid store already used src/ — only updated its test import of
indexeddb to use the new path.

All import paths, tsconfig includes, and biome script targets updated.
No logic changes.
