+++ packages/tapeworm/src/event_store_partition.ts
