+++ packages/tapeworm_persistence_store_indexeddb/src/idb_partition.ts
