+++ packages/tapeworm_persistence_store_indexeddb/tsconfig.json
@@ -8,5 +8,5 @@
     "lib": ["ES2022", "DOM"],
     "types": ["node"]
   },
-  "include": ["lib/**/*", "index.ts"]
+  "include": ["src/**/*", "index.ts"]
 }
