+++ packages/tapeworm_persistence_store_mongodb/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./lib";
+export { default } from "./src";
