+++ packages/tapeworm_persistence_store_mongodb/tsconfig.json
@@ -6,5 +6,5 @@
     "declaration": true,
     "ignoreDeprecations": "6.0"
   },
-  "include": ["lib/**/*", "index.ts"]
+  "include": ["src/**/*", "index.ts"]
 }
