+++ packages/tapeworm_persistence_store_remote/index.ts
@@ -1,1 +1,1 @@
-export { Client, Server } from "./lib";
+export { Client, Server } from "./src";
